requirements.txt
cvxopt==1.3.0
keras==2.10.0
matplotlib==3.5.2
numpy==1.22.4
pandas==1.4.2
python-dotenv==0.21.0
scipy==1.8.1
seaborn==0.12.1
tensorflow==2.10.0
tqdm==4.64.0
